Futures/CompletableFuturesv2.0/parallel-asynchronous-using-java/src/main/java/com/learnjava/parallelstreams/ParallelStreamExample.java
package com.learnjava.parallelstreams;

import com.learnjava.util.DataSet;

import java.util.List;
import java.util.stream.Collectors;

import static com.learnjava.util.CommonUtil.*;
import static com.learnjava.util.LoggerUtil.log;

public class ParallelStreamExample {

    private String addNameLengthTransform(String name) {
        delay(500);
        return name.length() + " - " + name;
    }

    public List<String> stringTransform(List<String> namesList) {
        return namesList.stream()
                .parallel()
                .map(this::addNameLengthTransform)
                .collect(Collectors.toList());
    }

    public static void main(String[] args) {
        List<String> namesList = DataSet.namesList();
        ParallelStreamExample parallelStreamExample = new ParallelStreamExample();
        startTimer();
        List<String> result = parallelStreamExample.stringTransform(namesList);
        timeTaken();
        log("Result: " + result);
    }
}
